fix: remove attachment if asked for

+++ R/mails.R
@@ -80,8 +80,7 @@ mailSend <- function(from = "RMail <laresbernardo@gmail.com>",
   stop_for_status(req)
   
   # Delete temporary files created
-  if (!is.null(file) & as != "attachment")
-    file.remove(file)
+  if (!is.null(file) & !"attachment" %in% as) file.remove(file)
   
   if (!quiet) message(paste(
     subject, "sent to", to, ifelse(!is.null(as), paste("with", as, "file"),"")))
